pyproject.toml
[project]
name = "ontology-data-agent"
version = "0.1.0"
description = "Ontology-driven data agent MVP"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.1",
    "langchain-core>=1.0",
    "rdflib>=7.0.0",
    "google-cloud-aiplatform>=1.100.0",
    "Faker>=28.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "streamlit>=1.30.0",
    "pandas>=2.0.0",
    "plotly>=5.0.0",
    "openai>=1.0.0",
    "langfuse>=2.0.0",
    "sqlglot>=25.0.0",
    "duckdb>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
